% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mh.R
\name{approximations}
\alias{approximations}
\alias{mvqt_pois_approx}
\alias{mvqt_binom_approx}
\title{Helper functions for approximations}
\usage{
mvqt_pois_approx(around, k, mean, Q)

mvqt_binom_approx(around, k, n, mean, Q)
}
\arguments{
\item{around}{The vector around which to perform an approximation.}

\item{k, n}{Number of observations}

\item{mean, Q}{Priors}
}
\description{
Helper functions for approximations
}
