% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mh.R, R/slice_sample.R
\name{mh_binom_reg}
\alias{mh_binom_reg}
\alias{binom_reg}
\alias{ss_binom_reg}
\title{Slice sample a binomial regression rate}
\usage{
mh_binom_reg(
  p,
  k,
  n,
  mean,
  precision,
  proposal = c("normal", "uniform", "quadratic taylor", "mv quadratic taylor"),
  ...,
  proposal_sd = 1
)

ss_binom_reg(
  p,
  k,
  n,
  mean,
  precision,
  ...,
  w = 1,
  nexpand = 10,
  ncontract = 100
)
}
\arguments{
\item{p}{the previous iteration of the logit-probability}

\item{k}{the realized value from the binomial distribution}

\item{n}{the number of trials}
}
\description{
Slice sample a binomial regression rate
}
\details{
\code{ss_binom_reg} slice samples and \code{mh_binom_reg} Metropolis-samples
  \code{p} conditional on \code{k}, \code{n}, \code{mean}, and \code{precision},
  where \code{k ~ Binom(n, expit(p))} and \code{p ~ N(mean, precision)}.

  In the case that \code{n} is zero, slice sampling is ignored in favor of a normal draw.
  In the special case when an entire (multivariate) row of \code{n} is zero, the entire row
  is simultaneously drawn (in R); when only some elements are zero, they're drawn univariately (in C++).
  Note that only the latter is used when \code{proposal} indicates multivariate MH.

  Both vectorized over \code{p}, \code{k}, \code{n}, and \code{mean}. If \code{precision} is a matrix,
  \code{p} is assumed to be multivariately distributed, and different internals are used.

  The internals are defined in C++.
}
