% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conjugacy.R
\name{conj_norm_mu}
\alias{conj_norm_mu}
\alias{conj_mvnorm_mu}
\alias{conj_matnorm_mu}
\alias{conj_lm_beta}
\alias{conj_matlm_beta}
\alias{conj_norm_tau}
\alias{conj_mvnorm_Q}
\alias{conj_matnorm_V}
\alias{conj_lm_tau}
\alias{conj_matlm_sigma}
\title{Draw Conjugate Samples}
\usage{
conj_norm_mu(y, tau, mu0 = 0, tau0 = 0.001, mult = 1)

conj_mvnorm_mu(y, Q, mu0 = rep_len(0, p), Q0 = diag(0.001, p), mult = 1)

conj_matnorm_mu(y, V, U = NULL, mu0, Q0)

conj_lm_beta(y, X, XtX = crossprod(X), tau, mu0, Q0)

conj_matlm_beta(y, X, V, U = NULL, mu0, Q0)

conj_norm_tau(y, mu, a0 = 0.001, b0 = 0.001)

conj_mvnorm_Q(y, mu, V0, v0, V0_inv = chol_inv(V0))

conj_matnorm_V(y, mu, U = NULL, V0, v0, V0_inv = chol_inv(V0))

conj_lm_tau(y, X, beta, Xbeta = X \%*\% beta, a0 = 0.001, b0 = 0.001)

conj_matlm_sigma(
  y,
  X,
  beta,
  Xbeta = X \%*\% beta,
  U = NULL,
  V0,
  v0,
  V0_inv = chol_inv(V0)
)
}
\arguments{
\item{y}{realizations from the distribution whose parameter is being drawn. For multivariate conjugacy,
this is an n-by-p matrix}

\item{mu0}{the prior mean of \code{mu} or \code{beta}.}

\item{mult}{An optional multiplier for the prior precision; useful in some cases}

\item{Q, tau}{the precision of the (multivariate) normal distribution from which \code{y} comes}

\item{Q0, tau0}{the prior precision of \code{mu}.}

\item{V, U}{the precision matrices for the matrix-normal distribution}

\item{X}{the data matrix on \code{beta}}

\item{XtX, V0_inv, Xbeta}{pre-computed "shortcut" arguments for efficiency reasons}

\item{mu}{the mean of the normal distribution from which \code{y} comes}

\item{a0, b0}{the parameters (shape and rate) of the gamma distribution prior on \code{tau}.}

\item{V0, v0}{the parameters (matrix and degrees of freedom) of the Wishart prior on \code{Q} or \code{V}}

\item{beta}{the coefficients on \code{X}}
}
\description{
Sample from the posterior (conditional on all other parameters) in the conjugate setting.
}
