% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{distributions}
\alias{distributions}
\alias{dmvnorm}
\alias{dmvlnorm}
\alias{dmvnorm_diff}
\alias{dmvlnorm_diff}
\alias{dnorm_diff}
\alias{dlnorm_diff}
\alias{dmatnorm}
\alias{dmatnorm_diff}
\alias{dlogitnorm}
\alias{rdirich}
\title{Statistical Distributions}
\usage{
dmvnorm(x, mu, Q, detQ = determinant(Q, logarithm = TRUE)$modulus, log = TRUE)

dmvlnorm(x, mu, Q, detQ = determinant(Q, logarithm = TRUE)$modulus, log = TRUE)

dmvnorm_diff(x, y, mu, Q, log = TRUE, byrow = FALSE)

dmvlnorm_diff(x, y, mu, Q, log = TRUE, byrow = FALSE)

dnorm_diff(x, y, mu, tau, log = TRUE, byrow = FALSE)

dlnorm_diff(x, y, mu, tau, log = TRUE, byrow = FALSE)

dmatnorm(
  x,
  mu,
  V,
  U = NULL,
  detV = determinant(V, logarithm = TRUE)$modulus,
  detU = determinant(U, logarithm = TRUE)$modulus,
  log = TRUE
)

dmatnorm_diff(x, y, mu, V, U = NULL, log = TRUE)

dlogitnorm(x, mu, sd, log = FALSE)

rdirich(n, alpha)
}
\arguments{
\item{x, y}{vector of quantiles}

\item{mu}{the mean}

\item{detQ, detU, detV}{Pre-computed log-determinants}

\item{log}{Should the log-density be returned?}

\item{byrow}{Should the densities be summed (FALSE, the default) or returned separately by row (TRUE).}

\item{tau, Q}{the precision (matrix)}

\item{V, U}{the precision matrices for the matrix-normal distribution}

\item{sd}{the standard deviation}

\item{n}{Number of deviates to produce.}

\item{alpha}{the Dirichlet parameter vector or matrix.}
}
\description{
Statistical Distributions
}
\details{
\code{dmvnorm_diff(x, y, mu, Q)} is equivalent to (but usually twice as fast as)
  \code{dmvnorm(x, mu, Q) - dmvnorm(y, mu, Q)}. Likewise \code{dmatnorm_diff}.
}
