% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_sample.R
\name{gibbs}
\alias{gibbs}
\alias{gs_matlm_beta}
\title{Draw Gibbs Samples}
\usage{
gs_matlm_beta(
  beta,
  ...,
  a0 = rep_len(-Inf, length(beta)),
  b0 = rep_len(Inf, length(beta)),
  params.only = FALSE
)
}
\arguments{
\item{beta}{the coefficients on \code{X}}

\item{...}{Passed through to \code{\link{conj_matlm_beta}}}

\item{a0, b0}{Optional limits of truncation for a truncated normal prior.}

\item{params.only}{Should just a list of the updated parameters be returned?}
}
\description{
Sample from the posterior conditional on all but one parameter (at a time).
}
