% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute_conj_mvnorm_mu}
\alias{impute_conj_mvnorm_mu}
\title{Impute Conjugate Values}
\usage{
impute_conj_mvnorm_mu(y, mu, impute, Q, mu0, tau0)
}
\arguments{
\item{y}{A matrix of values drawn from the multivariate normal distribution
whose mean we are to impute.}

\item{mu}{the means we wish to impute. Should be the same size as \code{y}.
No NA's are allowed; indices to impute are passed to \code{impute}. Instead,
usually values are initialized and allowed to update within the MCMC framework.}

\item{impute}{A logical matrix the size of \code{y}, indicating where to impute.}

\item{Q}{the precision of the multivariate normal distribution from which \code{y} comes.
Should be square, with dimension \code{ncol(y)}.}

\item{mu0, tau0}{The prior means/precisions for \code{mu}.
Can be a matrix the same size as \code{mu}, in which case
only the values corresponding to \code{impute == TRUE} are used, or a vector with length
\code{sum(impute)}.}
}
\description{
Impute Conjugate Values
}
