% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_sample.R
\name{multinom_reg}
\alias{multinom_reg}
\alias{ss_multinom_reg}
\title{Slice sample a multinomial regression rate}
\usage{
ss_multinom_reg(
  p,
  z,
  k,
  mean,
  precision,
  ...,
  w = 1,
  nexpand = 10,
  ncontract = 100
)
}
\arguments{
\item{p}{the previous iteration of the logit-probability, as an array of dimension \code{r x i x j},
where \code{r} is the number of realizations, \code{i} is the number of (correlated)
betas (that is, the dimension of \code{precision}), and \code{j} is the number of
classes to split between. It is expected that \code{p[, , 1] == 0}}

\item{z}{a matrix of zeros and ones. The zeros determine so-called "structural zeros": outcomes
which are not possible. This is assumed not to change over the first dimension (\code{r}) of \code{p}.
This is of dimension \code{i x j}.}

\item{k}{the realized value from the binomial distribution; the same size as \code{p} (\code{r x i x j}).}

\item{mean}{the prior mean for \code{p}. Must be either a scalar or an array of size \code{r x i x (j-1)}}

\item{precision}{an array of dimension \code{i x i x (j-1)}.
The \code{j}-dimension is assumed to be independent.}

\item{...}{Other arguments (not used)}

\item{w}{The width of each expansion (to the right and left each)}

\item{nexpand}{The maximum number of expansions (to the right and left each)}

\item{ncontract}{The maximum number of contractions. If this is exceeded, the original value is returned}
}
\description{
Slice sample a multinomial regression rate
}
\details{
\code{ss_binom_reg} slice samples and \code{mh_binom_reg} Metropolis-samples
  \code{p} conditional on \code{k}, \code{n}, \code{mean}, and \code{precision},
  where \code{k_{ri\*} ~ multinom(n, expit(p_{ri\*}))} and \code{p_{r\*j} ~ N(mean_{r\*j}, precision_{j})}.

  \code{precision} must be a matrix at this time, by which
  \code{p} is assumed to be multivariately distributed (over the second dimension).

  The internals are defined in C++.
}
\seealso{
\code{\link{ss_pois_reg}}, \code{\link{ss_binom_reg}}, \url{https://en.wikipedia.org/wiki/Slice_sampling}
}
