% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.R
\name{sample_binom_reg}
\alias{sample_binom_reg}
\title{Sample a binomial regression rate}
\usage{
sample_binom_reg(
  p,
  k,
  n,
  mean,
  precision,
  method = c("slice", "normal", "uniform", "quadratic taylor", "mv quadratic taylor"),
  ...,
  width = 1,
  nexpand = 10,
  ncontract = 100
)
}
\arguments{
\item{p}{the previous iteration of the logit-probability}

\item{k}{the realized value from the binomial distribution}

\item{n}{the number of trials}

\item{mean}{the prior mean}

\item{precision}{the prior precision}

\item{method}{The method to use to propose new values. \code{"normal"} proposes normal deviates from the current
value. \code{"uniform"} proposes uniform deviates. \code{"gamma"} does moment-matching on the mean and variance
to propose a conjugate gamma proposal. \code{"quadratic taylor"} proposes using a second-order Taylor
approximation of the log-density at the current value, which amounts to a normal proposal with mean (usually) not equal
to the current value.}

\item{...}{Other arguments (not used)}

\item{width}{For \code{"normal"} proposals, the standard deviation(s) of proposals. For \code{"uniform"} proposals,
the half-width of the uniform proposal interval. For \code{"slice"}, the width of each expansion (to the right and left each).
For \code{"gamma"} a scaling factor to increase the variance of the proposal.}

\item{nexpand}{The maximum number of expansions (to the right and left each)}

\item{ncontract}{The maximum number of contractions. If this is exceeded, the original value is returned}
}
\description{
Sample a binomial regression rate
}
\details{
This function samples \code{p} conditional on \code{k}, \code{n}, \code{mean}, and \code{precision},
  where \code{k ~ Binom(n, expit(p))} and \code{p ~ N(mean, precision)}.

  In the case that \code{n} is zero, sampling is ignored in favor of a normal draw.
  In the special case when an entire (multivariate) row of \code{n} is zero, the entire row
  is simultaneously drawn (in R); when only some elements are zero, they're drawn univariately (in C++).
  Note that only the latter is used when \code{proposal} indicates multivariate MH.

  This is vectorized over \code{p}, \code{k}, \code{n}, and \code{mean}. If \code{precision} is a matrix,
  \code{p} is assumed to be multivariately distributed, and different internals are used.

  The internals are defined in C++.
}
\seealso{
\url{https://en.wikipedia.org/wiki/Slice_sampling},
\url{https://en.wikipedia.org/wiki/Metropolis–Hastings_algorithm}, \url{https://arxiv.org/pdf/1308.0657.pdf}
}
