% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinom.R
\name{sample_multinom_reg}
\alias{sample_multinom_reg}
\title{Sample a multinomial regression rate}
\usage{
sample_multinom_reg(
  p,
  z,
  k,
  mean,
  precision,
  method = c("slice"),
  ref = c("first", "last"),
  ...,
  width = 1,
  nexpand = 10,
  ncontract = 100
)
}
\arguments{
\item{p}{the previous iteration of the logit-probability, as an array of dimension \code{r x i x j},
where \code{r} is the number of realizations, \code{i} is the number of (correlated)
betas (that is, the dimension of \code{precision}), and \code{j} is the number of
classes to split between. It is expected that \code{p[, , ref] == 0}}

\item{z}{a matrix of zeros and ones. The zeros determine so-called "structural zeros": outcomes
which are not possible. This is assumed not to change over the first dimension (\code{r}) of \code{p}.
This is of dimension \code{i x j}.}

\item{k}{the realized value from the binomial distribution; the same size as \code{p} (\code{r x i x j}).}

\item{mean}{the prior mean for \code{p}. Must be either a scalar or an array of size \code{r x i x (j-1)}}

\item{precision}{an array of dimension \code{i x i x (j-1)}.
The \code{j}-dimension is assumed to be independent.}

\item{method}{The method to use to propose new values. \code{"normal"} proposes normal deviates from the current
value. \code{"uniform"} proposes uniform deviates. \code{"quadratic taylor"} proposes using a second-order Taylor
approximation of the log-density at the current value, which amounts to a normal proposal with mean (usually) not equal
to the current value.}

\item{ref}{One of \code{"first"} or \code{"last"} or a numeric value, denoting which "column" (the third dimension)
 of \code{p} is the reference.
If \code{"first"}, then \code{mean} and \code{precision} map to the second through j-th elements of \code{p}.
If \code{"last"}, then \code{mean} and \code{precision} map to the first through (j-1)-th
elements of \code{p}.
If \code{2}, then \code{mean} and \code{precision} map to the first and third through j-th elements of \code{p}.
Etc.}

\item{...}{Other arguments (not used)}

\item{width}{For \code{"normal"} proposals, the standard deviation(s) of proposals. For \code{"uniform"} proposals,
the half-width of the uniform proposal interval. For \code{"slice"}, the width of each expansion (to the right and left each)}

\item{nexpand}{The maximum number of expansions (to the right and left each)}

\item{ncontract}{The maximum number of contractions. If this is exceeded, the original value is returned}
}
\description{
Sample a multinomial regression rate
}
\details{
The internals are defined in C++.

  In the case that \code{n} is zero or \code{z} is zero, slice sampling is ignored in favor of a normal draw.
}
\seealso{
\code{\link{sample_pois_reg}}, \code{\link{sample_binom_reg}}, \url{https://en.wikipedia.org/wiki/Slice_sampling}
}
