% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_sample.R
\name{ss_binom_reg}
\alias{ss_binom_reg}
\title{Slice sample a binomial regression rate}
\usage{
ss_binom_reg(
  p,
  k,
  n,
  mean,
  precision,
  ...,
  w = 1,
  nexpand = 10,
  ncontract = 100
)
}
\arguments{
\item{p}{the previous iteration of the logit-probability}

\item{k}{the realized value from the binomial distribution}

\item{n}{the number of trials}

\item{mean}{the prior mean}

\item{precision}{the prior precision}

\item{...}{Other arguments (not used)}

\item{w}{The width of each expansion (to the right and left each)}

\item{nexpand}{The maximum number of expansions (to the right and left each)}

\item{ncontract}{The maximum number of contractions. If this is exceeded, the original value is returned}
}
\description{
Slice sample a binomial regression rate
}
\details{
This function slice samples \code{p} conditional on \code{k}, \code{n}, \code{mean}, and \code{precision},
  where \code{k ~ Binom(n, expit(p))} and \code{p ~ N(mean, precision)}.

  This is vectorized over \code{p}, \code{k}, \code{n}, and \code{mean}. If \code{precision} is a matrix,
  \code{p} is assumed to be multivariately distributed, and a different function is used.

  The internals are defined in C++.
}
\seealso{
\code{\link{ss_pois_reg}}, \url{https://en.wikipedia.org/wiki/Slice_sampling}
}
