% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_sample.R
\name{ss_pois_reg}
\alias{ss_pois_reg}
\title{Slice sample a Poisson regression rate}
\usage{
ss_pois_reg(L, k, mean, precision, ..., w = 1, nexpand = 10, ncontract = 100)
}
\arguments{
\item{L}{the previous iteration of the log-rate}

\item{k}{the realized value from the Poisson distribution}

\item{mean}{the prior mean}

\item{precision}{the prior precision}

\item{...}{Other arguments (not used)}

\item{w}{The width of each expansion (to the right and left each)}

\item{nexpand}{The maximum number of expansions (to the right and left each)}

\item{ncontract}{The maximum number of contractions. If this is exceeded, the original value is returned}
}
\description{
Slice sample a Poisson regression rate
}
\details{
This function slice samples \code{L} conditional on \code{k}, \code{mean}, and \code{precision},
  where \code{k ~ Pois(exp(L))} and \code{L ~ N(mean, precision)}.

  This is vectorized over \code{L}, \code{k}, and \code{mean}. If \code{precision} is a matrix,
  \code{L} is assumed to be multivariately distributed, and a different function is used.

  The internals are defined in C++.
}
\seealso{
\code{\link{ss_binom_reg}}, \code{\link{ss_multinom_reg}}, \url{https://en.wikipedia.org/wiki/Slice_sampling}
}
