% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_sample.R
\name{ss_poisson_regression}
\alias{ss_poisson_regression}
\title{Slice sample a Poisson regression rate}
\usage{
ss_poisson_regression(L, k, mean, precision)
}
\arguments{
\item{L}{the previous iteration of the log-rate}

\item{k}{the realized value from the Poisson distribution}

\item{mean}{the prior mean}

\item{precision}{the prior precision}
}
\description{
Slice sample a Poisson regression rate
}
\details{
This function slice samples \code{L} conditional on \code{k}, \code{mean}, and \code{precision},
  where \code{k ~ Pois(exp(L))} and \code{L ~ N(mean, precision)}.

  This is vectorized over \code{L}, \code{k}, and \code{mean}. If \code{precsision} is a matrix,
  \code{L} is assumed to be multivariately distributed, and a different function is used.

  The internals are defined in C++.
}
\seealso{
\url{https://en.wikipedia.org/wiki/Slice_sampling}
}
