% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar1.R
\name{times_flat_ar1}
\alias{times_flat_ar1}
\title{Effecient multiplication in the AR(1) setting}
\usage{
times_flat_ar1(X, Pflat, Xt = t(X))
}
\arguments{
\item{X, Xt}{A matrix and its transpose}

\item{Pflat}{a flattened version of an AR(1) matrix. More details on this in a future version}
}
\value{
The result of \code{t(X) \%*\% P}.
}
\description{
Effecient multiplication in the AR(1) setting
}
