% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{my_inv}
\alias{my_inv}
\alias{logit}
\alias{expit}
\title{Other Utilities}
\usage{
my_inv(x)

logit(p)

expit(x)
}
\arguments{
\item{x}{A matrix for \code{my_inv} or a numeric vector for \code{expit}}

\item{p}{A numeric vector}
}
\description{
Other Utilities
}
