% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mh.R, R/slice_sample.R
\name{mh_binom_reg}
\alias{mh_binom_reg}
\alias{binom_reg}
\alias{ss_binom_reg}
\title{Slice sample a binomial regression rate}
\usage{
mh_binom_reg(
  p,
  k,
  n,
  mean,
  precision,
  proposal = c("normal", "uniform", "quadratic taylor"),
  ...,
  proposal_sd = 1
)

ss_binom_reg(
  p,
  k,
  n,
  mean,
  precision,
  ...,
  w = 1,
  nexpand = 10,
  ncontract = 100
)
}
\arguments{
\item{p}{the previous iteration of the logit-probability}

\item{k}{the realized value from the binomial distribution}

\item{n}{the number of trials}

\item{mean}{the prior mean}

\item{precision}{the prior precision}

\item{proposal}{The method to use to propose new values. \code{"normal"} proposes normal deviates from the current
value. \code{"uniform"} proposes uniform deviates. \code{"quadratic taylor"} proposes using a second-order Taylor
approximation of the log-density at the current value, which amounts to a normal proposal with mean (usually) not equal
to the current value.}

\item{...}{Other arguments (not used)}

\item{proposal_sd}{For \code{"normal"} proposals, the standard deviation(s) of proposals. For \code{"uniform"} proposals,
the half-width of the uniform proposal interval.}

\item{w}{The width of each expansion (to the right and left each)}

\item{nexpand}{The maximum number of expansions (to the right and left each)}

\item{ncontract}{The maximum number of contractions. If this is exceeded, the original value is returned}
}
\description{
Slice sample a binomial regression rate
}
\details{
\code{ss_binom_reg} slice samples and \code{mh_binom_reg} Metropolis-samples
  \code{p} conditional on \code{k}, \code{n}, \code{mean}, and \code{precision},
  where \code{k ~ Binom(n, expit(p))} and \code{p ~ N(mean, precision)}.

  In the case that \code{n} is zero, slice sampling is ignored in favor of a normal draw.
  In the special case when an entire (multivariate) row of \code{n} is zero, the entire row
  is simultaneously drawn; when only some elements are zero, they're drawn univariately.

  Both vectorized over \code{p}, \code{k}, \code{n}, and \code{mean}. If \code{precision} is a matrix,
  \code{p} is assumed to be multivariately distributed, and different internals are used.

  The internals are defined in C++.
}
\seealso{
\url{https://arxiv.org/pdf/1308.0657.pdf}

\code{\link{ss_pois_reg}}, \code{\link{ss_multinom_reg}}, \url{https://en.wikipedia.org/wiki/Slice_sampling},
\url{https://en.wikipedia.org/wiki/Metropolis–Hastings_algorithm}
}
