% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mh.R, R/slice_sample.R
\name{mh_binom_reg}
\alias{mh_binom_reg}
\alias{binom_reg}
\alias{ss_binom_reg}
\title{Slice sample a binomial regression rate}
\usage{
mh_binom_reg(p, k, n, mean, precision, ..., proposal_sd = 1)

ss_binom_reg(
  p,
  k,
  n,
  mean,
  precision,
  ...,
  w = 1,
  nexpand = 10,
  ncontract = 100
)
}
\arguments{
\item{p}{the previous iteration of the logit-probability}

\item{k}{the realized value from the binomial distribution}

\item{n}{the number of trials}

\item{mean}{the prior mean}

\item{precision}{the prior precision}

\item{...}{Other arguments (not used)}

\item{proposal_sd}{The standard deviation(s) of proposals}

\item{w}{The width of each expansion (to the right and left each)}

\item{nexpand}{The maximum number of expansions (to the right and left each)}

\item{ncontract}{The maximum number of contractions. If this is exceeded, the original value is returned}
}
\description{
Slice sample a binomial regression rate
}
\details{
\code{ss_binom_reg} slice samples and \code{mh_binom_reg} Metropolis-samples
  \code{p} conditional on \code{k}, \code{n}, \code{mean}, and \code{precision},
  where \code{k ~ Binom(n, expit(p))} and \code{p ~ N(mean, precision)}.

  Both vectorized over \code{p}, \code{k}, \code{n}, and \code{mean}. If \code{precision} is a matrix,
  \code{p} is assumed to be multivariately distributed, and different internals are used.

  The internals are defined in C++.
}
\seealso{
\code{\link{ss_pois_reg}}, \code{\link{ss_multinom_reg}}, \url{https://en.wikipedia.org/wiki/Slice_sampling},
\url{https://en.wikipedia.org/wiki/Metropolis–Hastings_algorithm}
}
