% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional.R
\name{cond_mvnorm}
\alias{cond_mvnorm}
\title{Draw Conditional Samples}
\usage{
cond_mvnorm(y, mu, Q, which, params.only = FALSE)
}
\arguments{
\item{y}{A vector of values drawn from the multivariate normal distribution. Element \code{which} is ignored.}

\item{mu}{the mean of the normal distribution from which \code{y} comes}

\item{Q}{the precision of the multivariate normal distribution from which \code{y} comes}

\item{which}{Which index to get the conditional distribution of}

\item{params.only}{Should just a list of the updated parameters be returned?}
}
\description{
Sample from a conditional distribution.
}
