% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conjugacy.R
\name{conjugacy}
\alias{conjugacy}
\alias{conj_norm_mu}
\alias{conj_mvnorm_mu}
\alias{conj_matnorm_mu}
\alias{conj_lm_beta}
\alias{conj_matlm_beta}
\alias{conj_norm_tau}
\alias{conj_mvnorm_Q}
\alias{conj_matnorm_V}
\alias{conj_lm_tau}
\alias{conj_binom_p}
\title{Draw Conjugate Samples}
\usage{
conj_norm_mu(y, tau, mu0 = 0, tau0 = 0.001, ..., mult = 1, params.only = FALSE)

conj_mvnorm_mu(
  y,
  Q,
  mu0 = NULL,
  Q0 = diag(0.001, p),
  ...,
  newQ = mult * Q0 + n * Q,
  newQ.chol = gu_chol(newQ),
  mult = 1,
  params.only = FALSE
)

conj_matnorm_mu(
  y,
  V,
  U = NULL,
  mu0 = NULL,
  Q0,
  ...,
  newQ = V \%x\% U + Q0,
  newQ.chol = gu_chol(newQ),
  diag = FALSE,
  zero = NULL,
  params.only = FALSE
)

conj_lm_beta(
  y,
  X,
  XtX = t(X) \%*\% X,
  tau,
  mu0 = NULL,
  Q0,
  ...,
  newQ = tau * XtX + Q0,
  newQ.chol = gu_chol(newQ),
  params.only = FALSE
)

conj_matlm_beta(
  y,
  X,
  V,
  U = NULL,
  mu0 = NULL,
  Q0,
  ...,
  XtU = if (is.null(U)) t(X) else t(X) \%*\% U,
  XtUX = XtU \%*\% X,
  newQ = V \%x\% XtUX + Q0,
  newQ.chol = gu_chol(newQ),
  diag = FALSE,
  zero = NULL,
  params.only = FALSE
)

conj_norm_tau(y, mu, a0 = 0.001, b0 = 0.001, params.only = FALSE)

conj_mvnorm_Q(y, mu = NULL, V0, v0, V0_inv = chol_inv(V0), params.only = FALSE)

conj_matnorm_V(
  y,
  mu = NULL,
  U = NULL,
  V0,
  v0,
  ...,
  ytUy = t(ymu) \%*\% U \%*\% ymu,
  V0_inv = chol_inv(V0),
  params.only = FALSE
)

conj_lm_tau(
  y,
  X,
  beta,
  Xbeta = X \%*\% beta,
  a0 = 0.001,
  b0 = 0.001,
  params.only = FALSE
)

conj_binom_p(k, n, a0 = 1, b0 = 1, params.only = FALSE)
}
\arguments{
\item{y}{realizations from the distribution whose parameter is being drawn. For multivariate conjugacy,
this is an n-by-p matrix}

\item{mu0}{the prior mean of \code{mu} or \code{beta}. The default (\code{NULL}) is the same as an appropriately-dimensioned
vector/matrix of all zeros, but a little faster.}

\item{...}{Other arguments. Examples include \code{verbose=}, \code{take.chol=}, and
\code{Rstruct=}.}

\item{mult}{An optional multiplier for the prior precision; useful in some cases}

\item{params.only}{Should just a list of the updated parameters be returned?}

\item{Q, tau}{the precision of the (multivariate) normal distribution from which \code{y} comes}

\item{Q0, tau0}{the prior precision of \code{mu}.}

\item{V, U}{the precision matrices for the matrix-normal distribution}

\item{diag}{If \code{TRUE}, both \code{V} and \code{Q0} are assumed (but not confirmed!) to be diagonal,
which can speed up the Cholesky decomposition up to 100x. Default \code{FALSE}.}

\item{zero}{A matrix of ones and zeros, the same size as the beta to sample. Zero indicates a structural zero in the beta.
In the event that this is specified, everything returned is of size \code{sum(zero)}.}

\item{X}{the data matrix on \code{beta}}

\item{XtX, XtU, XtUX, ytUy, V0_inv, Xbeta, newQ, newQ.chol}{pre-computed "shortcut" arguments for efficiency reasons}

\item{mu}{the mean of the normal distribution from which \code{y} comes}

\item{a0, b0}{the parameters (shape and rate) of the gamma distribution prior on \code{tau},
or the parameters (shape1 and shape2) of the beta distribution prior on \code{p}.}

\item{V0, v0}{the parameters (matrix and degrees of freedom) of the Wishart prior on \code{Q} or \code{V}}

\item{beta}{the coefficients on \code{X}}

\item{k}{The number of successes}

\item{n}{The number of trials}
}
\description{
Sample from the posterior (conditional on all other parameters) in the conjugate setting.
}
