% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conjugacy.R
\name{conj_norm_mu}
\alias{conj_norm_mu}
\alias{conj_mvnorm_mu}
\alias{conj_matnorm_mu}
\alias{conj_lm_beta}
\alias{conj_matlm_beta}
\alias{conj_diagmatlm_beta}
\alias{conj_norm_tau}
\alias{conj_mvnorm_Q}
\alias{conj_matnorm_V}
\alias{conj_lm_tau}
\alias{conj_matlm_sigma}
\alias{conj_binom_p}
\title{Draw Conjugate Samples}
\usage{
conj_norm_mu(y, tau, mu0 = 0, tau0 = 0.001, mult = 1, params.only = FALSE)

conj_mvnorm_mu(
  y,
  Q,
  mu0 = rep_len(0, p),
  Q0 = diag(0.001, p),
  newQ.inv = chol_inv(mult * Q0 + n * Q),
  A = t(chol(newQ.inv)),
  mult = 1,
  use.chol = FALSE,
  params.only = FALSE
)

conj_matnorm_mu(
  y,
  V,
  U = NULL,
  mu0,
  Q0,
  newQ.inv = chol_inv(V \%x\% U + Q0),
  A = t(chol(newQ.inv)),
  diag = FALSE,
  use.chol = FALSE,
  params.only = FALSE
)

conj_lm_beta(
  y,
  X,
  XtX = crossprod(X),
  tau,
  mu0,
  Q0,
  newQ.inv = chol_inv(tau * XtX + Q0),
  A = t(chol(newQ.inv)),
  use.chol = FALSE,
  params.only = FALSE
)

conj_matlm_beta(
  y,
  X,
  V,
  U = NULL,
  mu0,
  Q0,
  diag = FALSE,
  use.chol = FALSE,
  params.only = FALSE
)

conj_diagmatlm_beta(
  y,
  X,
  V,
  U = NULL,
  mu0,
  Q0,
  use.chol = FALSE,
  params.only = FALSE
)

conj_norm_tau(y, mu, a0 = 0.001, b0 = 0.001, params.only = FALSE)

conj_mvnorm_Q(y, mu, V0, v0, V0_inv = chol_inv(V0), params.only = FALSE)

conj_matnorm_V(
  y,
  mu,
  U = NULL,
  V0,
  v0,
  V0_inv = chol_inv(V0),
  params.only = FALSE
)

conj_lm_tau(
  y,
  X,
  beta,
  Xbeta = X \%*\% beta,
  a0 = 0.001,
  b0 = 0.001,
  params.only = FALSE
)

conj_matlm_sigma(
  y,
  X,
  beta,
  Xbeta = X \%*\% beta,
  U = NULL,
  V0,
  v0,
  V0_inv = chol_inv(V0),
  params.only = FALSE
)

conj_binom_p(k, n, a0 = 1, b0 = 1, params.only = FALSE)
}
\arguments{
\item{y}{realizations from the distribution whose parameter is being drawn. For multivariate conjugacy,
this is an n-by-p matrix}

\item{mu0}{the prior mean of \code{mu} or \code{beta}.}

\item{mult}{An optional multiplier for the prior precision; useful in some cases}

\item{params.only}{Should just a list of the updated parameters be returned?}

\item{Q, tau}{the precision of the (multivariate) normal distribution from which \code{y} comes}

\item{Q0, tau0}{the prior precision of \code{mu}.}

\item{use.chol}{Logical, determining whether to use \code{MASS::\link{mvrnorm}()} (\code{FALSE}, the default) or \code{\link{chol_mvrnorm}}.}

\item{V, U}{the precision matrices for the matrix-normal distribution}

\item{diag}{If \code{TRUE}, both \code{V} and \code{Q0} are assumed (but not confirmed!) to be diagonal,
which can speed up the Cholesky decomposition up to 100x. Default \code{FALSE}.}

\item{X}{the data matrix on \code{beta}}

\item{XtX, V0_inv, Xbeta, newQ.inv, A}{pre-computed "shortcut" arguments for efficiency reasons}

\item{mu}{the mean of the normal distribution from which \code{y} comes}

\item{a0, b0}{the parameters (shape and rate) of the gamma distribution prior on \code{tau},
or the parameters (shape1 and shape2) of the beta distribution prior on \code{p}.}

\item{V0, v0}{the parameters (matrix and degrees of freedom) of the Wishart prior on \code{Q} or \code{V}}

\item{beta}{the coefficients on \code{X}}

\item{k}{The number of successes}

\item{n}{The number of trials}
}
\description{
Sample from the posterior (conditional on all other parameters) in the conjugate setting.
}
