% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{distributions}
\alias{distributions}
\alias{dmvnorm}
\alias{dmatnorm}
\alias{dlogitnorm}
\title{Statistical Distributions}
\usage{
dmvnorm(x, mu, Q, log = TRUE, use_trace = TRUE)

dmatnorm(x, mu, V, U, log = TRUE)

dlogitnorm(x, mu, sd, log = FALSE)
}
\arguments{
\item{x}{vector of quantiles}

\item{mu}{the mean}

\item{Q}{the precision matrix}

\item{log}{Should the log-density be returned?}

\item{use_trace}{Should the matrix trace be used, or a for-loop?}

\item{V, U}{the precision matrices for the matrix-normal distribution}

\item{sd}{the standard deviation}
}
\description{
Statistical Distributions
}
