% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{distributions}
\alias{distributions}
\alias{dmvnorm}
\alias{dmatnorm}
\alias{dlogitnorm}
\title{Statistical Distributions}
\usage{
dmvnorm(x, mu, Q, detQ = determinant(Q, logarithm = TRUE)$modulus, log = TRUE)

dmatnorm(
  x,
  mu,
  V,
  U = NULL,
  detV = determinant(V, logarithm = TRUE)$modulus,
  detU = determinant(U, logarithm = TRUE)$modulus,
  log = TRUE
)

dlogitnorm(x, mu, sd, log = FALSE)
}
\arguments{
\item{x}{vector of quantiles}

\item{mu}{the mean}

\item{Q}{the precision matrix}

\item{detQ, detU, detV}{Pre-computed log-determinants}

\item{log}{Should the log-density be returned?}

\item{V, U}{the precision matrices for the matrix-normal distribution}

\item{sd}{the standard deviation}
}
\description{
Statistical Distributions
}
