% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{dmvnorm}
\alias{dmvnorm}
\alias{dmatnorm}
\alias{dlogitnorm}
\alias{rwish}
\title{Statistical Distributions}
\usage{
dmvnorm(x, mu, Q, log = TRUE)

dmatnorm(x, mu, V, U, log = TRUE)

dlogitnorm(x, mu, sd, log = FALSE)

rwish(V, v)
}
\arguments{
\item{x}{vector of quantiles}

\item{mu}{the mean}

\item{Q}{the precision matrix}

\item{log}{Should the log-density be returned?}

\item{V, U}{the precision matrices for the matrix-normal distribution, or (V) the parameter to the Wishart distribution}

\item{sd}{the standard deviation}

\item{v}{the degree-of-freedom parameter for the Wishart distribution}
}
\description{
Statistical Distributions
}
