% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_sample.R
\name{gibbs}
\alias{gibbs}
\alias{gs_diagmatlm_beta}
\title{Draw Gibbs Samples}
\usage{
gs_diagmatlm_beta(
  beta,
  y,
  X,
  V,
  U = NULL,
  mu0,
  Q0,
  a0 = rep_len(-Inf, length(beta)),
  b0 = rep_len(Inf, length(beta)),
  use.chol = FALSE,
  params.only = FALSE
)
}
\arguments{
\item{beta}{the coefficients on \code{X}}

\item{y}{realizations from the distribution whose parameter is being drawn. For multivariate conjugacy,
this is an n-by-p matrix}

\item{X}{the data matrix on \code{beta}}

\item{V}{the precision matrices for the matrix-normal distribution}

\item{U}{the precision matrices for the matrix-normal distribution}

\item{mu0}{the prior mean of \code{mu} or \code{beta}.}

\item{Q0}{the prior precision of \code{mu}.}

\item{a0, b0}{Optional limits of truncation for a truncated normal prior.}

\item{use.chol}{Logical, determining whether to use \code{MASS::\link{mvrnorm}()} (\code{FALSE}, the default) or \code{\link{chol_mvrnorm}}.}

\item{params.only}{Should just a list of the updated parameters be returned?}
}
\description{
Sample from the posterior conditional on all but one parameter (at a time).
}
