% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.R
\name{sample_binom_reg}
\alias{sample_binom_reg}
\title{Sample a binomial regression rate}
\usage{
sample_binom_reg(
  p,
  k,
  n,
  mean,
  precision,
  method = c("slice", "normal", "uniform", "quadratic taylor", "mv quadratic taylor",
    "mv ind quadratic taylor", "mv beta"),
  ...,
  width = 1,
  nexpand = 10,
  ncontract = 100,
  acceptance = c("MH", "LL only", "regardless")
)
}
\arguments{
\item{p}{the previous iteration of the logit-probability}

\item{k}{the realized value from the binomial distribution}

\item{n}{the number of trials}

\item{mean}{the prior mean}

\item{precision}{the prior precision}

\item{method}{The method to use to propose new values.
\itemize{
  \item{\code{"normal"} proposes normal deviates from the current value.}
  \item{\code{"uniform"} proposes uniform deviates.}
  \item{\code{"gamma"} (only for Poisson) does moment-matching on the mean and variance
to propose a conjugate gamma proposal.}
  \item{\code{"mv gamma"} (only for Poisson) does the same, but ignores off-diagonal elements of the precision matrix
in the proposal (for speed's sake; this may result in low acceptance rates when elements are highly correlated).}
  \item{\code{"mv beta"} (only for binomial, and still somewhat experimental) does moment matching on a mean and variance
   approximated by the log-normal distribution (instead of logit-normal, for which there is no closed form)
   to propose a conjugate beta proposal; here again we ignore off-diagonal elements of the precision matrix for the proposal.}
  \item{\code{"quadratic taylor"} proposes using a second-order Taylor approximation of the log-density at the current value,
   which amounts to a normal proposal with mean (usually) not equal to the current value.}
   \item{\code{"mv quadratic taylor"} does the same, but uses a multivariate normal approximation.}
   \item{\code{"mv ind quadratic taylor"} proposes using a similar Taylor approximation, but approximates the log-density around
   the mean, instead of the current value. Furthermore, like \code{"mv beta"} and \code{"mv gamma"}, it ignores off-diagonal
   elements of the precision matrix for the proposal, which again might yield low acceptance rates when elements are highly correlated.
   Both of these simplifications yield a significant speed boost.}
}
Note that \code{"mv gamma"}, \code{"mv beta"} and \code{"mv [ind ]quadratic taylor"} accept or reject an entire row at a time.}

\item{...}{Other arguments (not used)}

\item{width}{For \code{"normal"} proposals, the standard deviation(s) of proposals. For \code{"uniform"} proposals,
the half-width of the uniform proposal interval. For \code{"slice"}, the width of each expansion (to the right and left each).
For \code{"gamma"}, \code{"beta"}, and \code{"mv ind quadratic taylor"} a scaling factor to increase the variance of the proposal.}

\item{nexpand}{The maximum number of expansions (to the right and left each)}

\item{ncontract}{The maximum number of contractions. If this is exceeded, the original value is returned}

\item{acceptance}{What should be the criteria for acceptance? "MH" indicates the usual Metropolis-Hastings update. "LL only" ignores the proposal
densities but considers the log-likelihoods. "regardless" accepts no matter what. This is useful for testing, or for
when the method is a gamma, beta, or quadratic approximation, which can be hard to accept if the initial starting point is low-density.}
}
\description{
Sample a binomial regression rate
}
\details{
This function samples \code{p} conditional on \code{k}, \code{n}, \code{mean}, and \code{precision},
  where \code{k ~ Binom(n, expit(p))} and \code{p ~ N(mean, precision)}.

  In the case that \code{n} is zero, sampling is ignored in favor of a normal draw.
  In the special case when an entire (multivariate) row of \code{n} is zero, the entire row
  is simultaneously drawn (in R); when only some elements are zero, they're drawn univariately (in C++).
  Note that only the latter is used when \code{proposal} indicates multivariate MH.

  This is vectorized over \code{p}, \code{k}, \code{n}, and \code{mean}. If \code{precision} is a matrix,
  \code{p} is assumed to be multivariately distributed, and different internals are used.

  The internals are defined in C++.
}
\seealso{
\url{https://en.wikipedia.org/wiki/Slice_sampling},
\url{https://en.wikipedia.org/wiki/Metropolis–Hastings_algorithm}, \url{https://arxiv.org/pdf/1308.0657.pdf}
}
