% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinom.R
\name{sample_multinom_reg}
\alias{sample_multinom_reg}
\title{Sample a multinomial regression rate}
\usage{
sample_multinom_reg(
  p,
  z,
  k,
  mean,
  precision,
  method = c("slice", "normal", "uniform", "quadratic taylor",
    "mv ind quadratic taylor", "mv truncated exponential"),
  ref = c("first", "last"),
  ...,
  diag = all(precision[upper.tri(precision)] == 0),
  zmax = NULL,
  width = 1,
  nexpand = 10,
  ncontract = 100,
  acceptance = c("MH", "LL only", "regardless")
)
}
\arguments{
\item{p}{the previous iteration of the logit-probability, as an array of dimension \code{r x i x j},
where \code{r} is the number of realizations, \code{i} is the number of (correlated)
betas (that is, the dimension of \code{precision}), and \code{j} is the number of
classes to split between. It is expected that \code{p[, , ref] == 0}. This can also be "flattened" to a
matrix of dimension \code{r x (i x j)}.}

\item{z}{a matrix (or array) of zeros and ones. The zeros determine so-called "structural zeros": outcomes
which are not possible. If two-dimensional, it is assumed not to change over the first dimension (\code{r}) of \code{p}.
This is of dimension \code{i x j} or \code{r x i x j}.}

\item{k}{the realized value from the binomial distribution; the same size as \code{p}.}

\item{mean}{the prior mean for \code{p}. Must be either a scalar or an array of size \code{r x i x (j-1)} (when \code{p} is 3-D) or
a "flattened" matrix of the same size.}

\item{precision}{an array of dimension \code{i x i x (j-1)} or a "flattened" matrix of the same size.
The \code{j}-dimension is assumed to be independent.}

\item{method}{The method to use to propose new values.
\itemize{
  \item{\code{"normal"} proposes normal deviates from the current value.}
  \item{\code{"uniform"} proposes uniform deviates.}
  \item{\code{"gamma"} (only for Poisson) does moment-matching on the mean and variance
to propose a conjugate gamma proposal.}
  \item{\code{"mv gamma"} (only for Poisson) does the same, but ignores off-diagonal elements of the precision matrix
in the proposal (for speed's sake; this may result in low acceptance rates when elements are highly correlated).}
  \item{\code{"mv beta"} (only for binomial, and still somewhat experimental) does moment matching on a mean and variance
   approximated by the log-normal distribution (instead of logit-normal, for which there is no closed form)
   to propose a conjugate beta proposal; here again we ignore off-diagonal elements of the precision matrix for the proposal.}
  \item{\code{"quadratic taylor"} proposes using a second-order Taylor approximation of the log-density at the current value,
   which amounts to a normal proposal with mean (usually) not equal to the current value.}
   \item{\code{"mv quadratic taylor"} does the same, but uses a multivariate normal approximation.}
   \item{\code{"mv ind quadratic taylor"} proposes using a similar Taylor approximation, but approximates the log-density around
   the mean, instead of the current value. Furthermore, like \code{"mv beta"} and \code{"mv gamma"}, it ignores off-diagonal
   elements of the precision matrix for the proposal, which again might yield low acceptance rates when elements are highly correlated;
   in particular, it is not recommended for \code{sample_multinom_reg}, whose data are a priori correlated (unless you increase the
   variance a lot, in which case it works okay).
   Both of these simplifications yield a significant speed boost.}
   \item{\code{"mv truncated exponential"} proposes using a first-order Taylor approximation of the log-density at the current value,
   which amounts to a truncated exponential proposal.}
}
Note that \code{"mv gamma"}, \code{"mv beta"} and \code{"mv [ind ]quadratic taylor"} accept or reject an entire row at a time.}

\item{ref}{One of \code{"first"} or \code{"last"} or a numeric value, denoting which "column" (the third dimension)
 of \code{p} is the reference.
If \code{"first"}, then \code{mean} and \code{precision} map to the second through j-th elements of \code{p}.
If \code{"last"}, then \code{mean} and \code{precision} map to the first through (j-1)-th
elements of \code{p}.
If \code{2}, then \code{mean} and \code{precision} map to the first and third through j-th elements of \code{p}.
Etc.}

\item{...}{Other arguments (not used)}

\item{diag}{Logical, denoting whether the precision matrix is diagonal, for a small speed boost.}

\item{zmax}{For advanced use: a matrix, usually computed from the last two dimensions of \code{z}.}

\item{width}{For \code{"normal"} proposals, the standard deviation(s) of proposals. For \code{"uniform"} proposals,
the half-width of the uniform proposal interval. For \code{"slice"}, the width of each expansion (to the right and left each).
For \code{"gamma"}, \code{"beta"}, and \code{"mv ind quadratic taylor"} a scaling factor to increase the variance of the proposal
(it's squared, so that \code{width} is on the sd scale).}

\item{nexpand}{The maximum number of expansions (to the right and left each)}

\item{ncontract}{The maximum number of contractions. If this is exceeded, the original value is returned}

\item{acceptance}{What should be the criteria for acceptance? "MH" indicates the usual Metropolis-Hastings update. "LL only" ignores the proposal
densities but considers the log-likelihoods. "regardless" accepts no matter what. This is useful for testing, or for
when the method is a gamma, beta, or quadratic approximation, which can be hard to accept if the initial starting point is low-density.}
}
\description{
Sample a multinomial regression rate
}
\details{
The internals are defined in C++.

  In the case that \code{n} is zero or \code{z} is zero, slice sampling is ignored in favor of a
  (possibly multivarite) normal draw.
}
\seealso{
\code{\link{sample_pois_reg}}, \code{\link{sample_binom_reg}}, \url{https://en.wikipedia.org/wiki/Slice_sampling}
}
