% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois.R
\name{sample_pois_reg}
\alias{sample_pois_reg}
\title{Sample a Poisson regression rate}
\usage{
sample_pois_reg(
  L,
  k,
  mean,
  precision,
  method = c("slice", "normal", "uniform", "gamma", "quadratic taylor",
    "mv quadratic taylor"),
  ...,
  width = 1,
  nexpand = 10,
  ncontract = 100
)
}
\arguments{
\item{L}{the previous iteration of the log-rate}

\item{k}{the realized value from the Poisson distribution}

\item{mean}{the prior mean}

\item{precision}{the prior precision}

\item{method}{The method to use to propose new values. \code{"normal"} proposes normal deviates from the current
value. \code{"uniform"} proposes uniform deviates. \code{"gamma"} does moment-matching on the mean and variance
to propose a conjugate gamma proposal. \code{"quadratic taylor"} proposes using a second-order Taylor
approximation of the log-density at the current value, which amounts to a normal proposal with mean (usually) not equal
to the current value.}

\item{...}{Other arguments (not used)}

\item{width}{For \code{"normal"} proposals, the standard deviation(s) of proposals. For \code{"uniform"} proposals,
the half-width of the uniform proposal interval. For \code{"slice"}, the width of each expansion (to the right and left each).
For \code{"gamma"} a scaling factor to increase the variance of the proposal.}

\item{nexpand}{The maximum number of expansions (to the right and left each)}

\item{ncontract}{The maximum number of contractions. If this is exceeded, the original value is returned}
}
\description{
Sample a Poisson regression rate
}
\details{
This function samples \code{L} conditional on \code{k}, \code{mean}, and \code{precision},
  where \code{k ~ Pois(exp(L))} and \code{L ~ N(mean, precision)}.

  In the case that \code{k} is \code{NA} (akin to \code{\link{sample_binom_reg}} in the case \code{n == k == 0}),
  sampling is ignored in favor of a normal draw.
  In the special case when an entire (multivariate) row of \code{k} is \code{NA}, the entire row
  is simultaneously drawn (in R); when only some elements are \code{NA}, they're drawn univariately (in C++).
  Note that only the latter is used when \code{proposal} indicates multivariate MH.

  This is vectorized over \code{L}, \code{k}, and \code{mean}. If \code{precision} is a matrix,
  \code{L} is assumed to be multivariately distributed, and a different function is used.

  The internals are defined in C++.
}
\seealso{
\url{https://en.wikipedia.org/wiki/Slice_sampling},
\url{https://en.wikipedia.org/wiki/Metropolis–Hastings_algorithm}, \url{https://arxiv.org/pdf/1308.0657.pdf}
}
