% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utilities}
\alias{utilities}
\alias{gu_chol}
\alias{chol.gu_chol}
\alias{chol_inv}
\alias{chol_mvrnorm}
\alias{logit}
\alias{expit}
\title{Other Utilities}
\usage{
gu_chol(x, ...)

\method{chol}{gu_chol}(x, ...)

chol_inv(x)

chol_mvrnorm(
  n = 1,
  mu,
  Sigma,
  Precision,
  A = if (missing(Sigma)) backsolve(chol(Precision), diag(1, nrow(Precision))) else
    t(chol(Sigma))
)

logit(p)

expit(x)
}
\arguments{
\item{x}{A matrix for \code{chol_inv} or a numeric vector for \code{expit}}

\item{...}{Other arguments (not in use at this time)}

\item{n}{Number of samples to draw}

\item{mu, Sigma, Precision}{parameters for a multivariate normal distribution. One of \code{Sigma} and
\code{Precision} must be supplied.}

\item{A}{the "square root" by which we multiply the independent normals. Used only for speed.}

\item{p}{A numeric vector}
}
\description{
Other Utilities
}
\details{
Unlike \code{MASS::\link[MASS]{mvrnorm}()}, \code{chol_mvrnorm()} uses the Cholesky decomposition
  to draw random samples. This is usually faster; the largest gains are seen when \code{n} is small
  (when \code{n} is large, the matrix multiplication becomes the limiting factor).
}
